requirements.txt
behave==1.2.6
PyHamcrest==2.0.2
requests==2.24.0
selenium==3.141.0
PyPOM==2.2.0